适配微信7.0.6 以下功能暂时失效 1.发送gif表情包 2.获取用户地区

+++ app/src/main/java/com/gentcent/wechat/zzk/model/message/bean/MessageBean.java
@@ -46,7 +46,7 @@ public class MessageBean implements Serializable {
 	//消息ID
 	private String msgId;
 	//添加时间
-	private int addTime;
+	private long addTime;
 	//状态
 	private int status;
 	//文件大小
@@ -89,11 +89,11 @@ public class MessageBean implements Serializable {
 		this.msgId = msgId;
 	}
 	
-	public int getAddTime() {
+	public long getAddTime() {
 		return addTime;
 	}
 	
-	public void setAddTime(int addTime) {
+	public void setAddTime(long addTime) {
 		this.addTime = addTime;
 	}
